backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
black
isort
flake8
mypy
python-dotenv
pytest
